Move remaining C# text helpers from Metadata
Move FQN and member-selector normalization, operator notation, XML documentation comment parsing, and shared XML-doc text extraction into the dependency-free CSharpText leaf. Rewire Metadata, Services, Research, CLI, and tests without changing behavior.


Copilot-Session: bc2613de-a721-4efd-8390-56db684eefd7
